# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText API
  description: >
    API for the WASAText messaging application, which allows users to send and receive
    individual or group messages with support for text, photos, and reactions.
    Includes simplified login, user management, conversations, messages, and group handling.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
    description: Local Development Server
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: "user-id"
  schemas:
    ErrorResponse:
      type: object
      description: Standard error response payload.
      properties:
        error:
          type: string
          description: Human-readable error message describing what went wrong.
          example: "An error occurred."
          minLength: 1
          maxLength: 255
          pattern: '^.+$'
      required:
        - error
    User:
      type: object
      description: Represents an application user.
      properties:
        id:
          type: string
          description: Unique user identifier.
          example: "user123"
          minLength: 1
          maxLength: 64
          pattern: '^[A-Za-z0-9_-]+$'
        name:
          type: string
          description: Display name of the user.
          example: "Maria"
          minLength: 1
          maxLength: 50
          pattern: '^[A-Za-zÀ-ÿ0-9 .''-]{1,50}$'
        photo:
          type: string
          format: uri
          description: Public URL of the user profile photo.
          example: "https://example.com/maria.jpg"
          minLength: 1
          maxLength: 2048
          pattern: '^https?://.+$'
      required: [name]
    Conversation:
      type: object
      description: Represents a conversation (direct or group).
      properties:
        id:
          type: string
          description: Unique conversation identifier.
          example: "conv123"
          minLength: 1
          maxLength: 64
          pattern: '^[A-Za-z0-9_-]+$'
        type:
          type: string
          description: Type of conversation (direct or group).
          enum: [direct, group]
          example: "group"
          minLength: 5
          maxLength: 6
          pattern: '^(direct|group)$'
        name:
          type: string
          description: Human-readable name of the conversation (mainly for groups).
          example: "Work Friends"
          minLength: 1
          maxLength: 100
          pattern: '^[A-Za-zÀ-ÿ0-9 .''-]{1,100}$'
        photo:
          type: string
          format: uri
          description: Public URL of the conversation or group photo.
          example: "https://example.com/group-photo.jpg"
          minLength: 1
          maxLength: 2048
          pattern: '^https?://.+$'
        lastMessagePreview:
          type: string
          description: Short preview of the most recent message in the conversation.
          example: "See you tomorrow?"
          minLength: 0
          maxLength: 140
          pattern: '^[\s\S]*$'
        lastMessageAt:
          type: string
          format: date-time
          description: Timestamp of the most recent message in the conversation (ISO 8601).
          example: "2025-11-04T20:00:00Z"
        lastMessageSenderId:
          type: string
          description: ID of the user who sent the last message.
          example: "user123"
          minLength: 1
          maxLength: 64
          pattern: '^[A-Za-z0-9_-]+$'
        lastMessageSenderName:
          type: string
          description: Name of the user who sent the last message.
          example: "Juan"
          minLength: 1
          maxLength: 50
          pattern: '^[A-Za-zÀ-ÿ0-9 .''-]{1,50}$'
        lastMessageStatus:
          type: string
          description: Status of the last message.
          enum: [delivered, read]
          example: "read"
          minLength: 4
          maxLength: 9
          pattern: '^(delivered|read)$'
        unreadCount:
          type: integer
          description: Number of unread messages for the requesting user.
          example: 5
          minimum: 0
        participants:
          type: array
          description: List of users participating in the conversation.
          minItems: 1
          maxItems: 512
          items:
            $ref: "#/components/schemas/User"
      required: [id, type, participants]
...
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
paths:
  /session:
    post:
      tags:
        - Login
      summary: Logs in or registers a user
      description: >
        Checks if a user exists by name. If they exist, returns their ID.
        If not, creates a new user and returns the generated ID.
      operationId: doLogin
      requestBody:
        description: User name for login/registration.
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: The username to login with.
                  example: "Maria"
                  minLength: 3
                  maxLength: 16
              required:
                - name
      responses:
        '200':
          description: Login successful (User ID returned).
          content:
            application/json:
              schema:
                type: object
                properties:
                  identifier:
                    type: string
                    description: The User ID to use as the Bearer Token.
                    example: "550e8400-e29b-41d4-a716-446655440000"
                required:
                  - identifier
        '400':
          $ref: "#/components/responses/BadRequest"
        '500':
          $ref: "#/components/responses/InternalServerError"
  /me:
    get:
      tags:
        - Profile
      summary: Get current user
      description: Retrieve the profile of the authenticated user.
      operationId: doGetCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user profile.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        '401':
          description: Unauthorized. Authorization header is missing.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Unauthorized"
        '404':
          $ref: "#/components/responses/NotFound"
...
  /conversations:
    get:
      tags:
        - Conversations
      summary: List user conversations
      description: >
        Retrieve all conversations the user participates in.
        Results are sorted by most recent activity (last message timestamp or join date).
        For 'direct' conversations, the name and photo returned are those of the other participant.
      operationId: getMyConversations
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of conversations.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Conversation"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '500':
          $ref: "#/components/responses/InternalServerError"

    post:
      tags:
        - Groups
      summary: Create a new group
      description: Create a new group conversation. The authenticated user is automatically added as the creator.
      operationId: createGroup
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Payload for creating a group.
              properties:
                name:
                  type: string
                  description: Name of the new group.
                  example: "Work Friends"
                  minLength: 1
                  maxLength: 100
                  pattern: '^[A-Za-zÀ-ÿ0-9 .''-]{1,100}$'
                users:
                  type: array
                  description: List of user IDs to invite.
                  minItems: 0
                  maxItems: 511
                  items:
                    type: string
                    description: User identifier.
                    minLength: 1
                    maxLength: 64
                    pattern: '^[A-Za-z0-9_-]+$'
                  example: ["user123", "user456"]
              required:
                - name
                - users
      responses:
        '201':
          description: Group created successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Conversation"
        '400':
          $ref: "#/components/responses/BadRequest"
        '401':
          $ref: "#/components/responses/Unauthorized"
        '500':
          $ref: "#/components/responses/InternalServerError"
...